+++ app/src/androidTest/java/woowacourse/movie/view/ReservationActivityTest.kt
@@ -15,7 +15,7 @@ import woowacourse.movie.R
 import woowacourse.movie.domain.Minute
 import woowacourse.movie.domain.Movie
 import woowacourse.movie.view.mapper.toUiModel
-import woowacourse.movie.view.moviemain.reservation.ReservationActivity
+import woowacourse.movie.view.reservation.ReservationActivity
 import java.time.LocalDate
 
 @RunWith(AndroidJUnit4::class)
