+++ app/src/main/java/woowacourse/movie/view/reservation/ReservationContract.kt
@@ -1,4 +1,4 @@
-package woowacourse.movie.view.moviemain.reservation
+package woowacourse.movie.view.reservation
 
 import java.time.LocalDate
 import java.time.LocalTime
