+++ app/src/main/res/layout/activity_reservation.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:orientation="vertical"
-    tools:context=".view.moviemain.reservation.ReservationActivity">
+    tools:context=".view.reservation.ReservationActivity">
 
     <ScrollView
         android:id="@+id/movie_scrollview"
